app: fix privacy level handling
We are currently setting DEFAULT_PRIVACY_LEVEL to handle the
Project default privacy level. This has some drawbacks for
normal users as at import time they'll see projects already available
on the platform as addable when they are not.
On top of that normal users don't have a way to change the
privacy level of their projects.
On the other hand we are building all the versions, all
the branches in the repository, as public.
I don't think that's what we want. Instead we want to make the
projects public and the default version of the build private
so that users have control to enable or disable them.